+++ hadoop-hdds/interface-server/src/main/resources/proto.lock
@@ -806,31 +806,6 @@
             ]
           },
           {
-            "name": "StorageTypeProto",
-            "enum_fields": [
-              {
-                "name": "DISK",
-                "integer": 1
-              },
-              {
-                "name": "SSD",
-                "integer": 2
-              },
-              {
-                "name": "ARCHIVE",
-                "integer": 3
-              },
-              {
-                "name": "RAM_DISK",
-                "integer": 4
-              },
-              {
-                "name": "PROVIDED",
-                "integer": 5
-              }
-            ]
-          },
-          {
             "name": "ContainerReplicaProto.State",
             "enum_fields": [
               {
